+++ mycodo/mycodo_flask/utils/utils_settings.py
@@ -407,6 +407,9 @@ def settings_general_mod(form):
                 if form.brand_image.data:
                     mod_misc.brand_image = form.brand_image.data.read()
                 mod_misc.brand_image_height = form.brand_image_height.data
+                mod_misc.favicon_display = form.favicon_display.data
+                if form.brand_favicon.data:
+                    mod_misc.brand_favicon = form.brand_favicon.data.read()
                 mod_misc.daemon_debug_mode = form.daemon_debug_mode.data
                 mod_misc.hide_alert_success = form.hide_success.data
                 mod_misc.hide_alert_info = form.hide_info.data